Update launchpad.ahk
Cleaned code up a *little*, added verify flag and corresponding fourth line to config file. Saving config restarts dogelaunchpad. This is a means to get around the 'variable changes not implemented until application restart' bug - I haven't decided if this is hacky or not. either way, it makes testing easier for me.